more types of blocks

+++ include/renderer/renderer.h
@@ -44,7 +44,7 @@ namespace SymoCraft{
         // void draw3DModel(const glm::vec3& position, const glm::vec3& scale, float rotation, const VoxelVertex* vertices, int verticesLength);
         // void drawLine(const glm::vec3& start, const glm::vec3& end, const Style& style);
         // void drawBox(const glm::vec3& block_center_coord, const glm::vec3& size, const Style& style);
-        void GenerateBlockFrameData(const glm::ivec3 &block_center_coord);
+        void GenerateBlockFrameData(const glm::vec3 &block_center_coord);
 
         void ReportStatus();
